Flatten scope type list (#2628)
Brings our dfa compilation time from around 700ms down to about 320ms

Fixes #2614

I had forgot, but we actually do parse `<user.any_alphanumeric_key>` se
we can show the users spoken form in the tutorial. I've now utilize this
further and actually add the spoken forms for the glyph scope type in
the flattened list.

Note that this implementation is somewhat hacky on purpose. Basically I
didn't want to touch our csv parser without first talking to pokey since
a lot of these list and spoken forms are used in places like the cheat
sheet and the tutorial. What I'm instead doing is keeping all the
existing lists and then creating a new list that is a flattened version
of them. That way the sheet sheet and other places can still use the
individual list and we are only using this flattened larger list for the
actual scope type capture. This is probably something we want to revisit
later, but for now we're getting a huge boast in dfa compilation time
with no changes to the speakable grammar and that I think is a clear
win.



## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [x] I have not broken the cheatsheet
- [x] Run Talon grammar tests

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>