Replace `private.switchStatementSubject` with value scope (#2999)
* We're already using `value` for the collection in for-each statements:
-
https://github.com/cursorless-dev/cursorless/blob/d55c744091394d8840a17e04ef8b7d21f03068b3/data/fixtures/scopes/javascript.core/value.foreach.scope
* Most of the tree sitter parsers call this field `value`. The other
alternative that's commonly used is `condition`, but we're already using
that in switch cases and I want to avoid conflicts.
-
https://github.com/cursorless-dev/cursorless/blob/d55c744091394d8840a17e04ef8b7d21f03068b3/queries/javascript.core.scm#L614

Since this scope is private we're not going to do any migration or
deprecation. It will just be removed. The this pull requests adds
`value` scope support for the switch subject though.

With that said I'm going to leave this open for a week in case someone
has input.